Total refactor, folder for constant headers init
