+++ pch.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+/*
+This is the precompiled header file for the Zorb Game. This header file contains all the constants that are used throughout the game.
+Some of those constant functions and variables include the following:
+- Game States
+- Zorb Names
+- Zorb Appearances
+- ANSI Escape Sequences
+- Game Art
+*/
+
+#include "constants/gameDefs.hpp" // Include gameDefs.hpp for game definitions
+#include "constants/zUtility.hpp" // Include gameDefs.hpp for game definitions
+#include "constants/zArt.hpp" // Include gameDefs.hpp for game definitions
